requirements.txt
pandas
re
matplotlib.pyplot
matplotlib.font_manager
koreanize_matplotlib